Use Emotion's theme prop for theming
This allows for easy multisite coloring, and dark mode toggles and the likes. While I have nothing planned at this time, this was pretty much a useless change. The primary reasoning is to reflect the code style I use in all projects nowadays, since this project is sort of a showcase.