Put how-to articles in their own tag
Probably going to differentiate them a bit more in the future. Might make them more grayish in the list since they're less relevant.

My plan right now is to try to alternate between complex gists like the package security one, and how-tos which are much easier to write. I have a how-to lined up for next week, and am working on a couple of regular gists at the same time. I am aiming for a gist each week, and if that becomes too much trouble, switching to no less than a gist per month. I might delay publishing the next how-to gist so I can repost the package security one to try and get it more views.