Base all font-sizes on browser defaults & replace px with em/rems
Based on https://adrianroselli.com/2024/03/the-ultimate-ideal-bestest-base-font-size-that-everyone-is-keeping-a-secret-especially-chet.html - I am a big fan of using browser defaults, and it never occurred to me before that font-size also applied to that.

I chose ems where I wanted the size to be dependent on the parent (the default). I chose rems where I thought it should be independent of the parent.