Add optional republishedAt to gists
I rewrote "The security risks of front-end dependencies" to the point that it's a brand new article, but it felt weird just marking it as an update as that's more intended for smaller updates. Publishing a brand new article while I had already published this before seemed weird too, as it would come across as the same for readers since the spirit of the article remains unchanged. Instead I needed a third option; marking articles as republished. This also allows me to provide an explanation of what changed.

It might be this is the only article I'll ever use this on since most rewrites will also change the spirit of the article (as I assume they'll be caused by a change of opinion), but it does seem like a reasonable solution.