Update accessibility article: nitpickings
Slightly shorter a few parts.

I'm mostly looking at this because I don't really like this heavy emphasis on custom modals in the article. It takes a lot of reading time while the <dialog> mention should suffice. But yet, removing it or hiding it in a <details> would make the modals section weirdly short and feel like it omits something that a lot of apps still do manually.