Update EAA article: reword accessible UI intro and implement segmented control
Embedded the WCAG aside in the accessible UI and content section since it was previously before the first mention of WCAG, which was wrong.

Also made very minor tweaks to the WCAG requirements, but the diff makes that hard to see because I wrapped it all in a div for the position sticky.