Update EAA article: add applicable websites and apps
While I did see this before early when researching, I forgot about it and, after researching other sources that didn't mention it or somehow suggested it applied to all services, the assumption started to live it was for everything and I never again saw something that disputed the assumption until vytah mentioned it [on Reddit](https://old.reddit.com/r/programming/comments/1ihe0z0/the_european_accessibility_act_for_websites_and/mb3bwoi/). While I would love for this directive to apply to everything, I don't want to be misleading or worse, wrong. ;)

(funny detail: I asked ChatGPT for feedback on my article and it never pointed this out. So much for using an all-knowing LLM for reviews (not that a friend would have caught it if they hadn't read the directive as well))