Add text appear animation to homepage title
I went out of my way to make this progressive enhancement since it's such an important piece of text, especially for search engines.

I'm not really a fan of these kinds of text animations; they're fragile, especially across browsers. Firefox is having a fit with this. But I think it adds enough here to be worth it - kinda. I'll just try it for a bit, I can always revert it.