Reduce project tag fontsize
A tiny improvement to the structure of the projects, but I really need to rework it someday. I reckon the top two sections are already impressive enough, but this section is boring. It could do with more images, content hidden behind "read more" toggles and just generally better styling.

I also need to consider how to integrate the blog in there.