Simplify Figure: remove required href
Wrapping an image in a link is dumb. Just put the caption in a link instead when needed, leaving control to the consumer.

Also made the caption text smaller and gray, so when it's not in a link (ie it doesn't have an underline), the connection with the text afterwards doesn't look weird.