Fix inconsistent punctuation
I've converged on using the en dash with two spaces for breaking up sentences where a comma or semicolon don't suffice. I do like using the semicolon despite it being unpopular in other text, because it has a technical aspect to it that matches my content. It's a legitimate character that has a clear usage, so why not.